test files reorganized
